+++ src/main/java/domain/model/consequence/Consequence.java
@@ -1,4 +1,4 @@
-package domain.model;
+package domain.model.consequence;
 
 public class Consequence {
     private final String value;
