+++ src/main/java/view/ResultView.java
@@ -1,11 +1,18 @@
 package view;
 
-import domain.model.*;
+import domain.model.consequence.Consequence;
+import domain.model.consequence.Consequences;
+import domain.model.ladder.Direction;
+import domain.model.ladder.Ladder;
+import domain.model.ladder.Line;
+import domain.model.ladder.Result;
+import domain.model.participant.People;
+import domain.model.participant.Person;
 
 import java.util.Map;
 import java.util.stream.IntStream;
 
-import static domain.model.Direction.*;
+import static domain.model.ladder.Direction.*;
 
 public class ResultView {
     private static final int MAX_INTERVAL = 6;
