+++ src/test/java/domain/model/ConsequenceTest.java
@@ -1,5 +1,6 @@
 package domain.model;
 
+import domain.model.consequence.Consequence;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
