+++ src/test/java/domain/model/PeopleTest.java
@@ -1,5 +1,6 @@
 package domain.model;
 
+import domain.model.participant.People;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
