+++ src/test/java/domain/model/PersonTest.java
@@ -1,5 +1,6 @@
 package domain.model;
 
+import domain.model.participant.Person;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
