+++ src/test/java/domain/model/ResultTest.java
@@ -1,5 +1,9 @@
 package domain.model;
 
+import domain.model.consequence.Consequences;
+import domain.model.ladder.Result;
+import domain.model.participant.People;
+import domain.model.participant.Person;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
